% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls_rho.R
\name{gsl_nls_loss}
\alias{gsl_nls_loss}
\title{Loss functions with tunable parameters}
\usage{
gsl_nls_loss(
  rho = c("default", "huber", "barron", "bisquare", "welsh", "optimal", "hampel", "ggw",
    "lqq"),
  cc = NULL
)
}
\arguments{
\item{rho}{character loss function name, one of \code{"default", "huber", "barron", "absolute", "bisquare", "welsh", "optimal", "hampel", "ggw", "lqq"}.}

\item{cc}{numeric vector of loss function tuning parameters, with length depends on the chosen loss function.
If \code{NULL}, the default tuning parameters are returned.}
}
\value{
A \code{list} with two components:
\itemize{
\item rho
\item cc
}
with meanings as explained under 'Arguments'.
}
\description{
Allow the user to tune the coefficient(s) of the loss functions
available in \code{\link{gsl_nls}} and \code{\link{gsl_nls_large}}.
}
\examples{
## huber loss function with default tuning parameter
huber_loss()
}
\seealso{
\url{https://CRAN.R-project.org/package=robustbase/vignettes/psi_functions.pdf}
}
