% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls-methods.R
\name{nobs.gsl_nls}
\alias{nobs.gsl_nls}
\title{Extract the number of observations}
\usage{
\method{nobs}{gsl_nls}(object, ...)
}
\arguments{
\item{object}{An object inheriting from class \code{"gsl_nls"}.}

\item{...}{At present no optional arguments are used.}
}
\value{
Integer number of observations similar to \code{\link[stats]{nobs}}
}
\description{
Returns the number of \emph{observations} from a \code{"gsl_nls"} object.
}
\examples{
## data
set.seed(1)
x <- 1:10
y <- 1 + 0.5 * (1:10) + rnorm(10, sd = 0.1)
## model
obj <- gsl_nls(fn = y ~ a + b * x, data = data.frame(x = x, y = y), start = c(a = 0, b = 1))
nobs(obj)
}
\seealso{
\code{\link[stats]{nobs}}
}
