% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls-methods.R
\name{residuals.gsl_nls}
\alias{residuals.gsl_nls}
\title{Extract model residuals}
\usage{
\method{residuals}{gsl_nls}(object, type = c("response", "pearson"), ...)
}
\arguments{
\item{object}{An object inheriting from class \code{"gsl_nls"}.}

\item{type}{character; if \code{"response"} the raw residuals are returned, if \code{"pearson"}
the Pearson are returned, i.e. the raw residuals divided by their standard error.}

\item{...}{At present no optional arguments are used.}
}
\description{
Returns the model residuals from a fitted \code{"gsl_nls"} object. 
\code{resid} can also be used as an alias.
}
\examples{
## data
set.seed(1)
x <- 1:10
y <- 1 + 0.5 * (1:10) + rnorm(10, sd = 0.1)
## model
obj <- gsl_nls(fn = y ~ a + b * x, data = data.frame(x = x, y = y), start = c(a = 0, b = 1))
residuals(obj)
}
\seealso{
\code{\link[stats]{residuals}}
}
